+++ CHANGELOG.md
@@ -10,9 +10,13 @@ This change log follows the format documented in [Keep a CHANGELOG].
 
 ## [Unreleased]
 
-## [1.28.6] - 2017-06-12
+### Added
 
-- Hungarian language support added.
+- [Hungarian locale (hu)](https://github.com/date-fns/date-fns/pull/503)
+  (thanks to László Horváth [@horvathlg](https://github.com/horvathlg))
+
+- Add `step` to `eachDay` function. Thanks to [@BDav24](https://github.com/BDav24).
+  See PR [#487](https://github.com/date-fns/date-fns/pull/487).
 
 ## [1.28.5] - 2017-05-19